DocuMind-AI-master/requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
python-dotenv
google-generativeai
langchain-google-genai
sentence-transformers
